Handling possible errors in the command line arguments supplied, by defining custom exception classes.
Define 4 custom exception classes:
- InvalidArgument
- MissingArgument
- UnkownArgument
- InconsistentArguments

We use these classes to throw exceptions in the
ProcessCommandLine() function when the command line
arguments are not valid. The function now returns
not a bool, but a 'settings' structure that contains the
settings for the program according to the cmd arguments.

Also, minor update in the Caesar cipher class,
to catch exceptions when the key is not valid
[a condition for negative ('-') keys, and a return to a default value (0) have been added].