Throw exception InvalidKey when the keys supplied to the VigenereCipher and CaesarCipher are invalid.
This is done by checking if the key is empty, or contains
non-alphabetic characters for the VigenereCipher, or if the key is not a
positive integer for the CaesarCipher.
This is done by adding a new exception class InvalidKey to the
CustomExceptions.hpp file, and throwing this exception in the
VigenereCipher and CaesarCipher classes when the key is invalid.

Also add tests for the new exception in the testCaesarCipher and
testVigenereCipher files. The tests check that the exception is thrown
when an invalid key is supplied, and that the exception message is
correct.